Fix handle lifetimes to correctly represent the intended meaning
HandleOwners have a parent, and the underlying handle.
The handle has a borrow to nothing inside the HandleOwner. Nothing is
considered to be 'static.
When the handle is passed out by copy (yet to be implemented), the handle
carries has a borrow to the HandleOwner ('owner). This is especially used for
'parent lifetimes when creating new HandleOwners (such as
DeviceMemoryOwner borrowing &'parent DeviceOwner)