OMG, there are more problems than I thought
This is some work on making traits to better define the HandleOwner
system, and also make a Handles --HandleOwner-- array like collection.

Thre is still so much to do.
The traits now should be renamed to better reflect the important parts
of what they represent.
Also, we need to generalize Owners as maybe not actually owning the
handle such as the images returned by VkGetSwapchainImages.

Maybe HandleOwners should become HandleManagers<Owned/Borrowed>
where the generic type parameter represents if the manager actually
owns the handle (and should therefore drop it or not)

Just an Idea, but somthing needs to be done to the above end after this
commit