revert retrun types from HandleVec to Vec
In future, after GAT is available, I want to make the return container
generic and user configurable (with at least Vec and static Array options).

Also, it would be nice if I could remove the need to store commands and
parent handles with the handle owner at some point.

In any event, I do not think that HandleVec will be my way forward.

still keeping it in the code for now though as dead code