fix issue with example on linux
although KHR_external_fence is available,
there is an error when trying to load ONLY
KHR_external_fence, without KHR_external_fence_fd, for example.

This is probably a bug on my currently old
linux system, but I don't want these kinds of issues on other people
computers for now.

The use of the extensions was only to test the
context macros anyway, and no actual functionality was needed