Feature/Extension/Command trait reoganizing
ok I know this is a lot, and I can't even explain it here.
I am just a solo dev with no time to work on this away
from my full time job. I worked on this commit for weeks
and it is a lot of suff that needs to work together.

Hopefully this is the final way my command trait system
will work (knock on wood). I still need to make the Device
and further child dispatchable handles work with the new
system.

This also adds some initial p_next support, but I need to
make use of the new system to allow tagging the capabilities
of a PhysicalDevice after creation.

I should really spend some time to update the docs now. I
should really get a proper explination of the internals. I
just don't have time man.